overwrite original Anilist import to use V2
