+++ app/models/list_import/anilist/row.rb
@@ -1,98 +1,126 @@
+# frozen_string_literal: true
+
 class ListImport
   class Anilist
     class Row
-      attr_reader :node, :type
-
       def initialize(node, type)
         @node = node
-        @type = type # anime || manga
+        @type = type.to_s # anime | manga
       end
 
-      def media
-        key = "#{type}/#{media_info[:id]}"
+      def media_mapping
+        anilist_mapping = Mapping.lookup(anilist_key, media.id)
 
-        Mapping.lookup('anilist', key) ||
-          Mapping.guess(type.classify.safe_constantize, media_info)
-      end
+        return anilist_mapping if anilist_mapping.present?
 
-      def media_info
-        {
-          id: node[type]['id'],
-          title: node[type]['title_romaji'],
-          subtype: node[type]['type'],
-          episode_count: node[type]['total_episodes'],
-          chapter_count: node[type]['total_chapters']
-        }.compact
-      end
+        other_mapping = Mapping.lookup(mal_key, mal_id) ||
+                        Mapping.guess(type.classify.safe_constantize, media_info)
 
-      def status
-        case node['list_status']
-        when 'completed' then :completed
-        when /ing\z/ then :current
-        when /\Aplan/ then :planned
-        when 'on-hold' then :on_hold
-        when 'dropped' then :dropped
-        end
-      end
-
-      def progress
-        type == 'anime' ? node['episodes_watched'] : node['chapters_read']
-      end
-
-      def volumes
-        return unless type == 'manga'
-
-        node['volumes_read']
-      end
-
-      def rating
-        return nil if node['score_raw'].zero?
-
-        # 100-point scale to 20-point scale (raw)
-        [(node['score_raw'].to_f / 5).ceil, 2].max
-      end
-
-      def reconsume_count
-        type == 'anime' ? node['rewatched'] : node['reread']
-      end
-
-      def notes
-        node['notes']
-      end
-
-      def started_at
-        return nil if node['started_on'].nil?
-
-        date = node['started_on'].split('/')
-
-        DateTime.strptime(format_date(date), '%F')
-      rescue
-        nil
-      end
-
-      def finished_at
-        return nil if node['finished_on'].nil?
-
-        date = node['finished_on'].split('/')
-
-        DateTime.strptime(format_date(date), '%F')
-      rescue
-        nil
+        Mapping.create(item: other_mapping, external_site: anilist_key, external_id: media.id) if other_mapping.present?
+        other_mapping
       end
 
       def data
-        %i[status progress rating reconsume_count notes started_at finished_at]
-          .map { |k|
-            [k, send(k)]
-          }.to_h
+        fields.map { |field| [field, send(field)] }.to_h.compact
       end
 
       private
 
-      def format_date(date)
-        date[1] ||= '01'
-        date[2] ||= '01'
-        date.join('-')
+      attr_reader :node, :type
+
+      def fields
+        %i[
+          rating status reconsume_count progress
+          notes started_at finished_at
+        ]
+      end
+
+      # For mapping guess
+      def media_info
+        {
+          title: title,
+          subtype: type,
+          episode_count: media.episodes,
+          chapter_count: media.chapters
+        }.compact
+      end
+
+      # 100-point scale to 20-point scale (raw)
+      # rating -> score
+      def rating
+        return nil if node.score.zero?
+
+        [(node.score.to_f / 5).ceil, 2].max
+      end
+
+      def status
+        case node.status.downcase
+        when 'completed' then :completed
+        when 'current' then :current
+        when 'planning' then :planned
+        when 'paused' then :on_hold
+        when 'dropped' then :dropped
+        end
+      end
+
+      # reconsume_count -> repeat
+      def reconsume_count
+        node.repeat
+      end
+
+      def progress
+        node.progress
+      end
+
+      def notes
+        node.notes
+      end
+
+      # 2020-05-19 -> yyyy-mm-dd
+      def started_at
+        return if node.started_at.to_h.compact.blank?
+
+        # will properly convert to include 0 before days/months
+        formatted_date(node.started_at)
+      end
+
+      # finished_at -> completed_at
+      def finished_at
+        return if node.completed_at.to_h.compact.blank?
+
+        # will properly convert to include 0 before days/months
+        formatted_date(node.completed_at)
+      end
+
+      def formatted_date(date_node)
+        date_node.to_h.symbolize_keys.values_at(:year, :month, :day).join('-').to_date.to_s
+      end
+
+      def anilist_key
+        "AniList/#{type}"
+      end
+
+      def mal_key
+        "MyAnimeList #{type.capitalize}"
+      end
+
+      def media
+        node.media
+      end
+
+      def mal_id
+        media.id_mal
+      end
+
+      def titles
+        media.title
+      end
+
+      def title
+        titles.romaji.presence ||
+          titles.english.presence ||
+          titles.native.presence ||
+          titles.user_preferred
       end
     end
   end
