--- app/policies/list_import/anilist_v2_policy.rb
@@ -1,1 +1,0 @@
-class ListImport::AnilistV2Policy < ListImportPolicy; end
