--- spec/fixtures/list_import/anilist/access_token.json
@@ -1,6 +1,0 @@
-{
-  "access_token":"RahlCEqLEgc6GPu2zBYqiGgOy2aLLZEjYridJrCO",
-  "token_type":"Bearer",
-  "expires_in":3600,
-  "expires":1476497488
-}
