+++ spec/fixtures/list_import/anilist/anime_completed_accel_world.json
