+++ spec/fixtures/list_import/anilist/manga_current_arifureta.json
