--- spec/fixtures/list_import/anilist/nuck-anime.json
@@ -1,178 +1,0 @@
-{
-  "id": 26989,
-  "display_name": "Nuck",
-  "anime_time": 136,
-  "manga_chap": 0,
-  "about": "",
-  "list_order": 0,
-  "adult_content": true,
-  "forum_homepage": 0,
-  "legacy_lists": false,
-  "donator": 0,
-  "following": false,
-  "image_url_lge": "https://cdn.anilist.co/img/dir/user/reg/default.png",
-  "image_url_med": "https://cdn.anilist.co/img/dir/user/sml/default.png",
-  "image_url_banner": "https://cdn.anilist.co/img/dir/userbanner/",
-  "title_language": "romaji",
-  "score_type": 0,
-  "custom_list_anime": [],
-  "custom_list_manga": [],
-  "stats": {
-    "status_distribution": {
-      "anime": {
-        "watching": 0,
-        "plan to watch": 0,
-        "completed": 1,
-        "dropped": 0,
-        "on-hold": 0
-      },
-      "manga": {
-        "reading": 0,
-        "plan to read": 0,
-        "completed": 0,
-        "dropped": 0,
-        "on-hold": 0
-      }
-    },
-    "score_distribution": {
-      "anime": {
-        "10": 0,
-        "20": 0,
-        "30": 0,
-        "40": 0,
-        "50": 0,
-        "60": 0,
-        "70": 0,
-        "80": 0,
-        "90": 0,
-        "100": 0
-      },
-      "manga": {
-        "10": 0,
-        "20": 0,
-        "30": 0,
-        "40": 0,
-        "50": 0,
-        "60": 0,
-        "70": 0,
-        "80": 0,
-        "90": 0,
-        "100": 0
-      }
-    },
-    "favourite_genres": {
-      "Hentai": 1
-    }
-  },
-  "advanced_rating": true,
-  "advanced_rating_names": [],
-  "notifications": 0,
-  "lists": {
-    "completed": [
-      {
-        "record_id": 5828444,
-        "series_id": 1639,
-        "list_status": "completed",
-        "score": 0,
-        "score_raw": 0,
-        "episodes_watched": 1,
-        "chapters_read": 1,
-        "volumes_read": 0,
-        "rewatched": 3,
-        "reread": 3,
-        "priority": 0,
-        "private": 0,
-        "hidden_default": 0,
-        "notes": null,
-        "advanced_rating_scores": [],
-        "custom_lists": [],
-        "started_on": null,
-        "finished_on": null,
-        "added_time": "2014-12-01T04:18:11+09:00",
-        "updated_time": "2014-12-01T04:18:18+09:00",
-        "anime": {
-          "id": 1639,
-          "title_romaji": "Boku no Pico",
-          "title_english": "Boku no Pico",
-          "title_japanese": "ぼくのぴこ",
-          "type": "OVA",
-          "start_date_fuzzy": 20060907,
-          "end_date_fuzzy": 20060907,
-          "season": null,
-          "series_type": "anime",
-          "synonyms": [
-            "Boku no Piko",
-            "My Pico",
-            "Boku no Pico OVA 1 - My Pico"
-          ],
-          "genres": [
-            "Hentai"
-          ],
-          "adult": true,
-          "average_score": 45,
-          "popularity": 1527,
-          "updated_at": 1475459650,
-          "image_url_sml": "https://cdn.anilist.co/img/dir/anime/sml/1639.jpg",
-          "image_url_med": "https://cdn.anilist.co/img/dir/anime/med/1639.jpg",
-          "image_url_lge": "https://cdn.anilist.co/img/dir/anime/reg/1639.jpg",
-          "total_episodes": 1,
-          "airing_status": "finished airing"
-        }
-      }
-    ],
-    "on_hold": [
-      {
-        "record_id": 13295885,
-        "series_id": 1,
-        "list_status": "on-hold",
-        "score": 0,
-        "score_raw": 0,
-        "episodes_watched": 1,
-        "chapters_read": 1,
-        "volumes_read": 0,
-        "rewatched": 0,
-        "reread": 0,
-        "priority": 0,
-        "private": 0,
-        "hidden_default": 0,
-        "notes": null,
-        "advanced_rating_scores": [],
-        "custom_lists": null,
-        "started_on": null,
-        "finished_on": null,
-        "added_time": "2012-05-30T11:56:33+09:00",
-        "updated_time": "2012-05-30T11:56:33+09:00",
-        "anime": {
-          "id": 1,
-          "title_romaji": "COWBOY BEBOP",
-          "title_english": "Cowboy Bebop",
-          "title_japanese": "カウボーイビバップ",
-          "type": "TV",
-          "start_date_fuzzy": 19980403,
-          "end_date_fuzzy": 19990424,
-          "season": null,
-          "series_type": "anime",
-          "synonyms": [],
-          "genres": [
-            "Action",
-            "Adventure",
-            "Comedy",
-            "Drama",
-            "Sci-Fi"
-          ],
-          "adult": false,
-          "average_score": 86.3,
-          "popularity": 11458,
-          "updated_at": 1475459350,
-          "image_url_sml": "https://cdn.anilist.co/img/dir/anime/sml/1.jpg",
-          "image_url_med": "https://cdn.anilist.co/img/dir/anime/med/1.jpg",
-          "image_url_lge": "https://cdn.anilist.co/img/dir/anime/reg/1.jpg",
-          "total_episodes": 26,
-          "airing_status": "finished airing"
-        }
-      }
-    ]
-  },
-  "custom_lists": [],
-  "css": []
-}
